+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Don't let users manually add form widgets
 * Allow overriding alchemy-field title & description in extra widget settings
+* Allow overriding an alchemy-field's used wrapper-view
 
 ## 0.1.11 (2022-07-23)